refactor: extract row-actions helper and remove frontend dead code
Share buildStandardRowActions across admin list pages and drop unused PlaceholderPage plus duplicate useApi tests.
